Code as deployed at San Diego Maker Faire 2015
Moved the MIDI output to the MTP8 boards to Serial2.

The third MIDI adapter (on Serial3) is now connected to the Organelle
in both directions, instead of the level converters formerly connected
to Serial. This eliminates the need for debug/download mode switching.

Keyboard Note On and Note Off events are now relayed to the Organelle,
to allow it to interact with the user. This feature is controlled by a
new flag in the custom SysEx message, and defaults to on.

There is now only one external MIDI source (the third MIDI adapter,
which is connected to the Organelle), so separate handling of MIDI
messages from the Organelle (formerly connected to Serial) is no longer
included.